Test utills: prevent a warning from MSVC.

+++ tests/test_utils.cpp
@@ -43,7 +43,7 @@ std::string fileContents(const std::string &fileName)
 
 void printErrors(const libcellml::LoggerPtr &l, bool headings, bool kinds, bool rule)
 {
-    int width = floor(log10(l->errorCount()));
+    int width = int(floor(log10(l->errorCount())));
     for (size_t i = 0; i < l->errorCount(); ++i) {
         std::cout << "Error " << std::setw(width) << i + 1 << ": ";
         std::cout << l->error(i)->description();
